requirements.txt
# Requirements for Credit Card Transaction Pipeline

# Kafka
kafka-python==2.0.2

# Spark
pyspark==3.5.1

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Web Scraping & HTTP
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Airflow
apache-airflow==2.7.3
apache-airflow-providers-apache-spark==4.3.0
